pyproject.toml
[project]
name = "kltn"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.0",
    "torch>=2.2.0",
    "torchvision",
    "torchaudio",
    "tqdm>=4.66.5",
    "streamlit>=1.55.0",
    "wandb>=0.25.0",
    "python-dotenv>=1.2.2",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "scikit-learn>=1.8.0",
    "plotly>=6.6.0",
    "playwright>=1.58.0",
    "beautifulsoup4>=4.14.3",
    "polars>=1.39.3",
    "duckdb>=1.5.1",
]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"

[tool.uv.sources]
torch = { index = "pytorch-cu118" }
torchvision = { index = "pytorch-cu118" }
torchaudio = { index = "pytorch-cu118" }
